update PR according to comments. Functions were moved to Container or Image classes
